requirements.txt
Django==1.8
psycopg2==2.6.2
djangorestframework==3.4.5